Fix: To throw unchecked exception for HTTP status 404
For unchecked exceptions handling in handle404 method,
in most of the cases the handle.getReturnType is usually not
of type action response class. If it is of type ActionResponse class
the code has been changed to successfully return ActionResponse object.
It is good throw all unchecked exceptions if handle.getReturnType is not equal to
Action Response.